3.2.1 dev

See merge request inf/DolphinScheduler!16

+++ dolphinscheduler-dao/src/main/java/org/apache/dolphinscheduler/dao/entity/ProjectNode.java
@@ -60,7 +60,7 @@ public class ProjectNode {
     private Integer clusterCode;
 
     @TableField("data_source_code")
-    private Long dataSourceCode;
+    private Integer dataSourceCode;
 
     @TableField("node_id")
     private String nodeId;
